+++ src/components/QRcomponents/QRcreator/index.js
