+++ src/components/QRcomponents/QRform/index.js
